nl-query-app/backend/requirements.txt
Flask==2.0.1
werkzeug==2.0.1
flask-cors==3.0.10
psycopg2-binary==2.9.5
sentence-transformers==2.2.0
huggingface-hub==0.8.1
transformers==4.19.2
torch==1.12.0
numpy==1.23.0
scikit-learn==1.1.1
nltk==3.7
tqdm==4.64.0
pyyaml==6.0
regex==2022.6.2
requests==2.28.1
